改善: mainからcontentsにディレクトリ名を変更
