--- src/components/main/docs/DocsNav.astro
@@ -1,15 +1,0 @@
----
-import { docsList } from "@/components/main/docs/docsList";
-
-const docs = docsList();
----
-
-<div id="nav-docs" class="md:(flex flex-col) hidden min-w-64 border-(r-2 docs-secondary)">
-  <div class="flex flex-col gap-8 sticky top-20 p-8">
-    <div class="flex flex-col gap-2">
-      {docs.map((doc) => (
-        <a class="p-2 rounded-md text-docs-head outline-cyan-600 hover:bg-stone-300" href={doc.url}>{doc.frontmatter.header}</a>
-      ))}
-    </div>
-  </div>
-</div>
