--- src/components/main/pages/PageBox.astro
@@ -1,14 +1,0 @@
----
-import { formatDateToJP } from '@/utils/dateFormat';
-const { header, updated } = Astro.props;
-const updatedDate = formatDateToJP(updated);
----
-<div class="box-main max-w-2xl h-full">
-  <div class="sm:(flex justify-between) pb-2 border-b">
-    <h1 class="text-4xl font-500">{header}</h1>
-    <p class="mt-4">最終更新日 : {updatedDate}</p>
-  </div>
-  <div class="markdown mt-8">
-    <slot />
-  </div>
-</div>
