+++ src/layouts/docsLayout.astro
@@ -1,7 +1,7 @@
 ---
 import Layout from "./layout.astro";
-import DocsBox from "@/components/main/docs/DocsBox.astro";
-import DocsNav from "@/components/main/docs/DocsNav.astro";
+import DocsBox from "@/components/contents/docs/DocsBox.astro";
+import DocsNav from "@/components/contents/docs/DocsNav.astro";
 
 interface Props {
   frontmatter: {
