+++ src/layouts/pageLayout.astro
@@ -1,5 +1,5 @@
 ---
-import PageBox from '@/components/main/pages/PageBox.astro';
+import PageBox from '@/components/contents/pages/PageBox.astro';
 import Layout from './layout.astro';
 
 interface Props {
